+++ react/components/Collapsible.js
@@ -4,14 +4,21 @@ import withScrollAnimation from './ScrollAnimation'
 import { Border } from './Helpers'
 import useDevice from '../hooks/useDevice'
 
-const Collapsible = ({ children, top, leanMode, animation }) => {
+const Collapsible = ({
+  children,
+  leanMode,
+  animation,
+  didAnimate,
+  onAnimate,
+}) => {
   const { desktop } = useDevice()
+  onAnimate(didAnimate)
 
   return desktop && !leanMode ? (
     <div
-      className={`${animation} fixed z-2 w-100 bg-base`}
+      className={`${animation} relative bg-base`}
       style={{
-        top: top,
+        zIndex: -1, // Animate under fixed content
         willChange: 'transform', // Better performance when animating
       }}
     >
@@ -19,18 +26,16 @@ const Collapsible = ({ children, top, leanMode, animation }) => {
       <Border />
     </div>
   ) : (
-    <Border fixed top={top} />
+    <Border />
   )
 }
 
 Collapsible.propTypes = {
   animation: PropTypes.string.isRequired,
   leanMode: PropTypes.bool,
-  top: PropTypes.number,
 }
 
 Collapsible.defaultProps = {
-  top: 0,
   leanMode: false,
 }